AKR(Frontend): Fix filtering by authorisation status logic.
If a translator has both an effective and an expired authorisation for the same language pair,
do not classify that translator as having an expired authorisation (at least, on the basis of
said authorisations). If there are additional expired authorisations *without* corresponding
effective authorisations (again, for the same language pair), then the translator shall of course
be classified as having expired authorisations.