AKR(Frontend): Adjust test data to ensure fix is tested.
Added an extra expired authorisation with same language pair as an effective authorisation for translator.
As the expected counts of translators matching different authorisation statuses is left unchanged,
this effectively tests that the filtering logic works as expected.

The case that a translator actually *does* have expired authorisations is already covered by the test data.