OTR(Frontend): Unset box-shadow on PublicInterpreterListingRow.
Trying to make the box-shadow work well when the CollapsibleRow is expanded proved to be too challenging.
Likewise, trying to make a single row occupy two rows' worth space (ie. setting rowSpan to 2)
when clicking on the row didn't seem to work; the table layout seemed to explode when attempting that.